Fix aligning extra line outside selection
When selection whole lines, a line directly beneath the current selection would also get aligned, even though its not part of the selection. This is because the selection range is a line and column index, so the selection.end corresponds to the next line (after the selection) but column 0, rather than the last line/index of the selection.

Fix is to manually check if the selection ends with a column index of 0, and if so we -1 from the line index.